# docs/swagger.yaml
basePath: /
definitions:
...
  fiber.Map:
    additionalProperties: true
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Appointment:
    properties:
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/models.User'
      customer_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_recurring:
        type: boolean
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      recur_pattern:
        allOf:
        - $ref: '#/definitions/models.Recurrence'
        description: E.g., "weekly", "monthly", "daily"
      recur_pattern_id:
        type: integer
      service:
        $ref: '#/definitions/models.Service'
      service_id:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.AppointmentStatus'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.AppointmentStatus:
    enum:
    - pending
    - confirmed
    - canceled
    - completed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusCanceled
    - StatusCompleted
  models.BusinessDetails:
    properties:
      address:
        type: string
      business_hours:
        type: string
      business_license:
        type: string
      business_name:
        type: string
      certificate_urls:
        type: string
      city:
        type: string
      country:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      profile_picture_url:
        type: string
      provider_id:
        type: integer
      state:
        type: string
      tax_number:
        type: string
      updatedAt:
        type: string
      website:
        type: string
      zip_code:
        type: string
    type: object
  models.DayOfWeek:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - Sunday
    - Monday
    - Tuesday
    - Wednesday
    - Thursday
    - Friday
    - Saturday
  models.Permission:
    properties:
      action:
        description: e.g., "create", "read", "update", "delete"
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        description: e.g., "appointments", "users", etc.
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    type: object
...
  models.Recurrence:
    properties:
      appointment_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_after:
        description: Number of occurrences
        type: integer
      frequency:
        description: '"daily", "weekly", "monthly"'
        type: string
      id:
        type: integer
      next_run:
        type: string
      updatedAt:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.Service:
    properties:
      buffer_time:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: Time between appointments
      cost:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      discount:
        description: Discount percentage
        type: number
      discounted_price:
        type: number
      duration:
        $ref: '#/definitions/time.Duration'
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      created_at:
        type: string
      customer_appointments:
        items:
          $ref: '#/definitions/models.Appointment'
        type: array
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      name:
        type: string
      otp:
        type: string
      otp_expires_at:
        type: string
      password:
        type: string
      provided_services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      working_hours:
        items:
          $ref: '#/definitions/models.WorkingHours'
        type: array
    type: object
...
  models.WorkingHours:
    properties:
      break_end:
        description: Optional break end time
        type: string
      break_start:
        description: Optional break start time
        type: string
      createdAt:
        type: string
      day_of_week:
        $ref: '#/definitions/models.DayOfWeek'
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: Format "HH:MM" in 24h
        type: string
      id:
        type: integer
      provider:
        $ref: '#/definitions/models.User'
      provider_id:
        type: integer
      start_time:
        description: Format "HH:MM" in 24h
        type: string
      updatedAt:
        type: string
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
...
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the Appointment App.
  title: Appointment App API
  version: "1.0"
paths:
...
  /providers:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all service providers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of providers with pagination
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                limit:
                  type: integer
                page:
                  type: integer
                pages:
                  type: integer
                providers:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
                total:
                  type: integer
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all providers
      tags:
      - providers
  /providers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details for a specific provider including working hours
        and business details
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Provider details
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                business_details:
                  $ref: '#/definitions/models.BusinessDetails'
                provider:
                  $ref: '#/definitions/models.User'
              type: object
        "404":
          description: Provider not found
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get provider details
      tags:
      - providers
...
  /providers/{id}/services:
    get:
      consumes:
      - application/json
      description: Retrieve services offered by a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of services
          schema:
            items:
              $ref: '#/definitions/models.Service'
            type: array
        "404":
          description: Provider not found
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get provider services
      tags:
      - providers
  /providers/available-time-slots/{provider_id}:
    get:
      consumes:
      - application/json
      description: Retrieve available appointment slots for a provider on a given
        date
      parameters:
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: integer
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        required: true
        type: string
      - description: Service ID
        in: query
        name: service_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Available slots
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                date:
                  type: string
                provider_id:
                  type: string
                service_id:
                  type: string
                slots:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Invalid date format or missing service ID
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Service not found or does not belong to provider
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get available appointment slots
      tags:
      - providers
  /providers/category/{categoryId}:
    get:
      consumes:
      - application/json
      description: Retrieve providers that offer services in a specific category
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of providers
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                count:
                  type: integer
                providers:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get providers by category
      tags:
      - providers
  /providers/featured:
    get:
      consumes:
      - application/json
      description: Retrieve a list of featured or top-rated providers
      produces:
      - application/json
      responses:
        "200":
          description: List of providers
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                providers:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get featured providers
      tags:
      - providers
  /providers/nearby:
    get:
      consumes:
      - application/json
      description: Retrieve providers near the user's location based on latitude,
        longitude, and radius
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: string
      - description: Longitude
        in: query
        name: lng
        required: true
        type: string
      - default: "10"
        description: Radius in kilometers
        in: query
        name: radius
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of nearby providers
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                lat:
                  type: string
                lng:
                  type: string
                providers:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
                radius:
                  type: string
              type: object
        "400":
          description: Location parameters (lat, lng) are required
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get nearby providers
      tags:
      - providers
  /providers/search/service:
    get:
      consumes:
      - application/json
      description: Search for providers by name, business name, or service
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of providers
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                count:
                  type: integer
                providers:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "400":
          description: Search query is required
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/fiber.Map'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Search providers
      tags:
      - providers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"